featues(Sales Invoice): Add sales invoice api post

+++ models/accounting_models/accounting_so_dr_details_view_model.go
@@ -10,7 +10,7 @@ type SalesOrderWithDRDetails struct {
 	UnitPrice      float64 `json:"unit_price"`
 	UomName        string  `json:"uom_name"`
 	TotalCost      float64 `json:"total_cost"`
-	Discount       uint    `json:"discount"`
+	Discount       float32 `json:"discount"`
 	Status         string  `json:"status"`
 	DeliveryDate   string  `json:"delivery_date"`
 }
